test: move transport config.json tests into config-async.vitest.ts (#549)
Final-review found tests/config-transport-file.vitest.ts raced
tests/config-async.vitest.ts over the shared root config.json fixture
(createConfigFileManager hard-codes a single path + backup, and vitest
runs files in parallel). config-async.vitest.ts already owns that
fixture, so the four transport normalization cases move there instead,
reusing its existing partial logger mock (now also stubbing
logTransportConfigWarning). The standalone file is removed.
